+++ webman/app/controller/system/SysBannerController.php
@@ -4,6 +4,7 @@ namespace app\controller\system;
 
 use support\Request;
 use app\service\SysBannerService;
+use app\service\PermissionService;
 use app\common\AjaxResult;
 use app\common\TableDataInfo;
 
@@ -11,6 +12,9 @@ class SysBannerController
 {
     public function list(Request $request)
     {
+        if (PermissionService::lacksPermi($request->loginUser, 'system:banner:list')) {
+            return json(['code' => 403, 'msg' => '没有操作权限']);
+        }
         $params = convert_to_snake_case($request->all());
         $result = SysBannerService::selectBannerList($params);
         return TableDataInfo::result($result->items(), $result->total());
@@ -18,6 +22,9 @@ class SysBannerController
 
     public function getInfo(Request $request)
     {
+        if (PermissionService::lacksPermi($request->loginUser, 'system:banner:list')) {
+            return json(['code' => 403, 'msg' => '没有操作权限']);
+        }
         $parts = explode('/', $request->path());
         $bannerId = intval(end($parts));
         $banner = SysBannerService::selectBannerById($bannerId);
@@ -29,6 +36,9 @@ class SysBannerController
 
     public function add(Request $request)
     {
+        if (PermissionService::lacksPermi($request->loginUser, 'system:banner:add')) {
+            return json(['code' => 403, 'msg' => '没有操作权限']);
+        }
         $data = convert_to_snake_case($request->post());
         $data['create_by'] = $request->loginUser->user->user_name ?? '';
         return AjaxResult::toAjax(SysBannerService::insertBanner($data) ? 1 : 0);
@@ -36,6 +46,9 @@ class SysBannerController
 
     public function edit(Request $request)
     {
+        if (PermissionService::lacksPermi($request->loginUser, 'system:banner:edit')) {
+            return json(['code' => 403, 'msg' => '没有操作权限']);
+        }
         $data = convert_to_snake_case($request->post());
         $data['update_by'] = $request->loginUser->user->user_name ?? '';
         return AjaxResult::toAjax(SysBannerService::updateBanner($data) ? 1 : 0);
@@ -43,6 +56,9 @@ class SysBannerController
 
     public function remove(Request $request)
     {
+        if (PermissionService::lacksPermi($request->loginUser, 'system:banner:remove')) {
+            return json(['code' => 403, 'msg' => '没有操作权限']);
+        }
         $bannerIds = explode(',', $request->input('bannerIds', ''));
         $bannerIds = array_map('intval', array_filter($bannerIds));
         return AjaxResult::toAjax(SysBannerService::deleteBannerByIds($bannerIds) ? 1 : 0);
